# backend/openapi.yaml
openapi: 3.0.3
info:
  title: NERP API
  version: 1.0.0
  description: API for sending commands from the React frontend to the Express backend.
servers:
  - url: http://localhost:3000
    description: Local development server
tags:
  - name: Health
    description: Server status checks
  - name: Commands
    description: Command records stored in PostgreSQL
  - name: Claude
    description: Claude AI prompt and response endpoints
  - name: Gemini
    description: Google Gemini creative endpoints (Atlas mentor, Paper Prism, Genesis Forge)
paths:
  /api/health:
    get:
      tags:
        - Health
      summary: Check API health
      responses:
        "200":
          description: API is running
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"
              example:
                status: ok
                service: nerp-api
  /api/commands:
    get:
      tags:
        - Commands
      summary: List recent commands
      responses:
        "200":
          description: Recent commands
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Command"
    post:
      tags:
        - Commands
      summary: Send a command
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCommandRequest"
            example:
              command: restart-worker
      responses:
        "201":
          description: Command created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Command"
        "400":
          description: Missing command value
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: Command is required.
  /api/claude:
    post:
      tags:
        - Claude
      summary: Ask Claude a question
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClaudeRequest"
            example:
              question: What is the NERP stack?
      responses:
        "200":
          description: Claude response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClaudeResponse"
        "400":
          description: Missing question value
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: Question is required.
        "500":
          description: Missing API key or server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: Missing ANTHROPIC_API_KEY. Add your Claude API key to backend/.env.
  /api/gemini/atlas:
    post:
      tags:
        - Gemini
      summary: Atlas Mentor — personalized advice with semantic weather
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AtlasRequest"
      responses:
        "200":
          description: Atlas reply (structured prose sections)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AtlasResponse"
        "400":
          description: Missing message or oversized payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Missing Gemini API key or server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: Missing GEMINI_API_KEY. Add your Gemini API key to backend/.env.
  /api/gemini/paper-prism:
    post:
      tags:
        - Gemini
      summary: Paper Prism — structured paper synthesis with cross-domain bridge
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaperPrismRequest"
      responses:
        "200":
          description: Structured prism analysis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaperPrismResponse"
        "400":
          description: Missing paper text or oversized payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Missing Gemini API key or server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: Missing GEMINI_API_KEY. Add your Gemini API key to backend/.env.
  /api/gemini/forge:
    post:
      tags:
        - Gemini
      summary: Genesis Forge — linked multi-modal creative outputs from one brief
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgeRequest"
      responses:
        "200":
          description: Linked creative bundle
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForgeResponse"
        "400":
          description: Missing brief or oversized payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Missing Gemini API key or server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: Missing GEMINI_API_KEY. Add your Gemini API key to backend/.env.
components:
  schemas:
    HealthResponse:
      type: object
      required:
        - status
        - service
      properties:
        status:
          type: string
          example: ok
        service:
          type: string
          example: nerp-api
    CreateCommandRequest:
      type: object
      required:
        - command
      properties:
        command:
          type: string
          minLength: 1
          example: restart-worker
    Command:
      type: object
      required:
        - id
        - command
        - status
        - created_at
      properties:
        id:
          type: integer
          example: 1
        command:
          type: string
          example: restart-worker
        status:
          type: string
          example: sent
        created_at:
          type: string
          format: date-time
          example: "2026-05-09T17:00:00.000Z"
    ClaudeRequest:
      type: object
      required:
        - question
      properties:
        question:
          type: string
          minLength: 1
          example: What is the NERP stack?
    ClaudeResponse:
      type: object
      required:
        - answer
        - model
      properties:
        answer:
          type: string
          example: NERP is a Node, Express, React, and PostgreSQL stack.
        model:
          type: string
          example: claude-sonnet-4-20250514
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          example: Something went wrong on the server.
    AtlasRequest:
      type: object
      required:
        - message
      properties:
        displayName:
          type: string
          example: Quinn
        coreSignal:
          type: string
          description: Short note about values, constraints, or goals used to personalize tone.
        message:
          type: string
          minLength: 1
        history:
          type: array
          description: Recent turns for continuity (client-managed).
          items:
            type: object
            required:
              - role
              - text
            properties:
              role:
                type: string
                enum:
                  - user
                  - model
              text:
                type: string
    AtlasResponse:
      type: object
      required:
        - answer
      properties:
        answer:
          type: string
    PaperPrismRequest:
      type: object
      required:
        - paperText
      properties:
        paperText:
          type: string
          minLength: 1
        focusArea:
          type: string
          description: Optional lens (methods, ethics, limitations, etc.).
    PaperPrismResponse:
      type: object
      required:
        - elevatorPitch
        - methodologyInFiveBullets
        - crossDomainBridge
        - skepticLens
        - methodAsPseudocode
        - listeningScore
      properties:
        elevatorPitch:
          type: string
        methodologyInFiveBullets:
          type: array
          items:
            type: string
        crossDomainBridge:
          type: string
        skepticLens:
          type: string
        methodAsPseudocode:
          type: string
        listeningScore:
          type: string
    ForgeRequest:
      type: object
      required:
        - brief
      properties:
        brief:
          type: string
          minLength: 1
        modalities:
          type: array
          description: >-
            Subset of python, dialogue, chords, ascii, shell. Empty means the model picks two surprises.
          items:
            type: string
            enum:
              - python
              - dialogue
              - chords
              - ascii
              - shell
    ForgeResponse:
      type: object
      required:
        - motif
        - pythonUtility
        - dialogueBeat
        - chordRoadmap
        - asciiArtifact
        - shellMicroscript
      properties:
        motif:
          type: string
        pythonUtility:
          type: string
        dialogueBeat:
          type: string
        chordRoadmap:
          type: string
        asciiArtifact:
          type: string
        shellMicroscript:
          type: string